Resolve qualified members in any position in the inheritance chain
This changes how the internal scope for contracts is laid out. Instead of
linking everything (state vars, functions, parent bases) to the lexical scope,
this is now done to an internal node, which then links to the lexical scope. The
lexical scope also contains the path to apply any dynamic scope changes set in
the contract (eg. any `using` directives).