Remove the generic type resolution for arrays and hard-code built-in in rules
The generic typing doesn't make much sense in Solidity since it doesn't have
generics. But some of the built-ins act as if those existed. Case in point:
`array.push()`. But using jump to scope to resolve the array's element type
conflicts with our other use of jump scopes which is extension scopes. Thus,
this commit removes the special handling for arrays (and therefore structs in
general) and instead hard-codes the `push()` built-in in the rules file.